2022/SWE/PlayerMatching/requirements.txt
requests
python-dotenv
flask
psycopg2-binary
Flask-SQLAlchemy==2.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
flask_bcrypt
twilio==7.7.1